Remove unused MyMap class, as well as some unused imports
Also remove the shebang line at the top of the file, since the file
defines no main function.

The Dockerfile for this project is using basemap version 1.1.0, which
is from May of 2017.  There is no newer official release, and the
library is showing its age.  For instance, basemap/__init__.py
references pyproj.pyproj_datadir, which causes an AttributeError to be
raised because there is no such attribute.  The simplest solution is
just to remove the unused MyMap class, which is the only place where
the basemap library is used.