Set up the Dockerfile for cross-compilation
Use the tonistiigi/xx Docker image to set up for cross-compilation.
This will improve image building speeds as the image will use the image
for the native platform but build for the target platform. This
eliminates the speed penalty of QEMU emulating non-native platforms and
compiling Python extensions in an emulated environment.